fact-check-ai_backend/requirements.txt
fastapi==0.110.0
pydantic==2.6.4
uvicorn==0.29.0
pymongo==4.6.3
newsapi-python==0.2.7
firebase-admin
pydantic[email]
python-multipart
llama-index
llama-index-llms-perplexity
llama-index-core
sqlalchemy>=2.0
tenacity
